refactor: improve dx
