+++ packages/app-store/templates/audit-log-implementation/lib/utils.ts
@@ -1,55 +1,7 @@
-import type { AuditLogEvent } from "@calcom/features/audit-logs/types";
-import { WEBAPP_URL } from "@calcom/lib/constants";
-import type { IconName } from "@calcom/ui";
-
-import type { GenericAuditLogClient } from "./AuditLogManager";
-
-export function getHref(
-  baseURL: string,
-  activeSettingsOption: { credentialId: string; activeOption: string }
-) {
-  const baseUrlParsed = new URL(baseURL, WEBAPP_URL);
-  baseUrlParsed.searchParams.set(activeSettingsOption.credentialId, activeSettingsOption.activeOption);
-  return baseUrlParsed.toString();
-}
-
-type AppSettingOptionEntry = {
-  name: string;
-  href: string;
-  icon: IconName;
-}[];
+import { getDefaultAppSettings } from "@calcom/features/audit-logs/constants";
+import type { AppSettingOptionEntry } from "@calcom/features/audit-logs/types";
 
 export function getAppSettingsOptions(credentialId: number): AppSettingOptionEntry[] {
-  return [
-    {
-      name: "Credentials",
-      href: "/apps/installed/auditLogs",
-      icon: "bar-chart",
-    },
-    {
-      name: "Logs",
-      href: getHref("/apps/installed/auditLogs", {
-        credentialId: credentialId.toString(),
-        activeOption: "logs",
-      }),
-      icon: "bar-chart",
-    },
-  ];
-}
-
-export function getGenericAuditLogClient(
-  apiKey: string,
-  projectId: string,
-  endpoint: string
-): GenericAuditLogClient {
-  return {
-    credentials: {
-      apiKey: apiKey,
-      projectId: projectId,
-      endpoint: endpoint,
-    },
-    reportEvent: (event: AuditLogEvent) => {
-      console.log({ event });
-    },
-  };
+  const defaultAppSettings = getDefaultAppSettings(credentialId);
+  return [...defaultAppSettings];
 }
